# live-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OxyLoans Live API",
    "version": "1.0.0",
    "description": "Live API gateway for OxyLoans e-commerce"
  },
  "servers": [
    {
      "url": "https://meta.oxyloans.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "summary": "Get trending products",
        "description": "Fetch all trending products with details",
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array"
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cart/{customer_id}": {
      "get": {
        "summary": "Get user cart",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Cart information"
          }
        }
      }
    },
    "/user/{user_id}": {
      "get": {
        "summary": "Get user profile",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile"
          }
        }
      }
    },
    "/offers": {
      "get": {
        "summary": "Get active offers",
        "responses": {
          "200": {
            "description": "Active offers"
          }
        }
      }
    },
    "/cart/add": {
      "post": {
        "summary": "Add item to cart",
        "parameters": [
          {
            "name": "customer_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item added"
          }
        }
      }
    },
    "/cart/remove": {
      "post": {
        "summary": "Remove item from cart",
        "parameters": [
          {
            "name": "customer_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed"
          }
        }
      }
    }
  }
}